requirements.txt
Flask==1.1.2
Pillow==7.1.2
PyMySQL==0.9.3
Werkzeug==1.0.1
cryptography==2.9.2
